refactor(ast): split ast.rs into module directory (closes #246)
Mechanical split of the 1,498-line ast.rs into six topical files under
ast/ for easier navigation. No logic changes; lib.rs still does
`pub use ast::*` so the public API is unchanged.

Layout:
- mod.rs   — ArenaVec, is_false helper, module re-exports
- names.rs — Name, NameKind, BuiltinType, TypeHint, TypeHintKind
- exprs.rs — Expr, ExprKind, NameStr, and expression sub-types
- stmts.rs — Stmt, StmtKind, control flow, namespace/use/const
- decls.rs — FunctionDecl, Param, ClassDecl and OOP sub-types
- misc.rs  — Program, Comment, CommentKind, Attribute, Arg
